Convert Int to BigInt

+++ compiler/lib/src/main/scala/analysis/Semantics/Param.scala
@@ -22,7 +22,7 @@ final case class Param(
 
 object Param {
 
-  type Id = Int
+  type Id = BigInt
 
   /** Creates a parameter from a parameter specifier
    *  Returns the new default opcode */
